+++ apps/mobile/app/(auth)/sign-in.tsx
@@ -5,7 +5,8 @@ import { router } from 'expo-router';
 import { Button } from 'react-native-paper';
 import * as Yup from 'yup';
 import { SignInDTO } from '@/types';
-import { PasswordInput, Surface, FieldInput, ParallaxScrollView, Text, HelloWave } from '@/components/common';
+import { Surface, ParallaxScrollView, Text, HelloWave } from '@/components/common';
+import { PasswordInput, FieldInput } from '@/components/form';
 import { useAuth } from '@/context/AuthContext';
 import { globalStyles } from '@/assets/styles';
