+++ apps/mobile/components/common/Text.tsx
@@ -14,11 +14,11 @@ export default function Text({ children, type = 'default', ...props }: CustomTex
       case 'link':
         return 'primary' as keyof MD3Colors;
       case 'secondary':
-        return 'secondary' as keyof MD3Colors;
+        return 'onSurfaceVariant' as keyof MD3Colors;
       case 'success':
         return 'tertiary' as keyof MD3Colors;
       case 'warning':
-        return 'tertiaryContainer' as keyof MD3Colors;
+        return 'onSecondaryContainer' as keyof MD3Colors;
       case 'danger':
         return 'error' as keyof MD3Colors;
       default:
