+++ apps/mobile/components/common/index.ts
@@ -1,10 +1,7 @@
-export { default as CaptchaInput } from './CaptchaInput';
 export { default as Collapsible } from './Collapsible';
 export { default as ExternalLink } from './ExternalLink';
-export { default as FieldInput } from './FieldInput';
 export { default as HelloWave } from './HelloWave';
 export { default as ParallaxScrollView } from './ParallaxScrollView';
-export { default as PasswordInput } from './PasswordInput';
 export { default as Surface } from './Surface';
 export { default as SafeAreaSurface } from './SafeAreaSurface';
 export { default as Carousel } from './Carousel';
