+++ apps/mobile/components/form/PasswordInput.tsx
@@ -1,9 +1,9 @@
 import { useField } from 'formik';
 import { useTranslation } from 'react-i18next';
-import { HelperText, TextInput, TextInputProps } from 'react-native-paper';
+import { TextInput, TextInputProps } from 'react-native-paper';
 import { useState } from 'react';
 import { Keyboard } from 'react-native';
-import Surface from './Surface';
+import Surface from '../common/Surface';
 
 interface PasswordInputProps extends TextInputProps {
   i18nKey: string;
@@ -11,8 +11,7 @@ interface PasswordInputProps extends TextInputProps {
 }
 
 export default function PasswordInput({ i18nKey, name, ...textInputProps }: PasswordInputProps) {
-  const [field, meta, helpers] = useField(name);
-  const hasError = meta.touched && Boolean(meta.error);
+  const [field, , helpers] = useField(name);
   const { t } = useTranslation();
   const [showPassword, setShowPassword] = useState(false);
 
@@ -27,15 +26,19 @@ export default function PasswordInput({ i18nKey, name, ...textInputProps }: Pass
         mode="outlined"
         label={t(`${i18nKey}.${name}`)}
         value={field.value}
-        error={hasError}
         placeholder={`${t('common.enter')}${t(`${i18nKey}.${name}`)}...`}
         onChangeText={(value) => helpers.setValue(value)}
         onBlur={() => helpers.setTouched(true)}
         secureTextEntry={!showPassword}
-        right={<TextInput.Icon icon={showPassword ? 'eye-off' : 'eye'} onPress={handleTogglePassword} />}
+        right={
+          <TextInput.Icon
+            icon={showPassword ? 'eye-off' : 'eye'}
+            onPress={handleTogglePassword}
+            style={{ opacity: field.value ? 1 : 0 }}
+          />
+        }
         {...textInputProps}
       />
-      {hasError && <HelperText type="error">{meta.error}</HelperText>}
     </Surface>
   );
 }
