+++ apps/mobile/components/form/index.ts
@@ -1,0 +1,3 @@
+export { default as FieldInput } from './FieldInput';
+export { default as PasswordInput } from './PasswordInput';
+export { default as CaptchaInput } from './CaptchaInput';
